requirements.txt
langgraph>=0.0.30
langchain-google-genai>=1.0.1
pydantic>=2.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
python-dotenv>=1.0.0
